created the util package module pin generation which creates the client id and secret key
