--- Auth/src/main/java/com/couponize/Auth/service/PinGenerator.java
@@ -1,18 +1,0 @@
-package com.couponize.Auth.service;
-
-
-import java.util.UUID;
-
-public class PinGenerator {
-
-    public static String IdGenerator() {
-        UUID applicationKey = UUID.randomUUID();
-        return applicationKey.toString();
-    }
-
-    public static String secretkey() {
-        UUID secretKey = UUID.randomUUID();
-        return secretKey.toString();
-    }
-
-}
